Do not close connection on pre-handshake command
tpafd would log and close the connection in case a client would issue
a non-hello protocol command before the hello commmand transaction had
finished.

While this is allowed in the protocol, the protocol also permits the
server to only terminate the particular command transaction that
violated the no-command-before-hello rule.

To more closely mimic pafd behavior and to be compatible with the
libpaf test suite, the more graceful option is the better one for
tpafd.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>